Implement web scraper to retrieve card details
I want to be able to build an app that makes it easy to see multiple
public transport card values in one place. This scraper will allow me
to pull data about my Myki cards so that I can use it inside the app.

It doesn't support Myki pass because I don't live in Melbourne and
therefore don't have a reason to buy bulk transport. It would be a
fairly trivial addition but I don't need it.