Sync scraped web data to Firebase (#7)
I want to be able to access the data all the time so I'm syncing it
to Firebase. This script will run on Heroku with a scheduler to
update it every 10mins or so. This is probably overkill :D

I considered running the whole app on Heroku but I don't want to wait
30secs for it to spin up each time and I don't want to pay for it. I
also considered doing the scraping from a cloud function in Firebase
but they don't allow you to ping external urls on the free version.
Again I didn't want to pay for it :D